Fixes bugged input field
In order to prevent the dragging of the canvas and some other adverse default behaviors, I had previously used an event listener on the body of the html and attached preventDefault to it. This made the input field unusable. The fix makes a check on the id of the elements that are clicked and only applies the preventDefault behavior if the element is not the input box.

Also made minor changes in the comments to clear up some functions functionality.